removing wf_convert mod
