+++ src/lib.rs
@@ -8,7 +8,6 @@ mod wf_account;
 mod wf_auth;
 mod wf_buffer;
 mod wf_codec;
-mod wf_convert;
 #[allow(dead_code)]
 mod wf_core;
 mod wf_crypto;
